+++ packages/datastore-storage-adapter/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.2.1...@aws-amplify/datastore-storage-adapter@2.2.2) (2025-12-10)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
+## [2.2.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.2.0...@aws-amplify/datastore-storage-adapter@2.2.1) (2025-11-06)
+
+**Note:** Version bump only for package @aws-amplify/datastore-storage-adapter
+
 # [2.2.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/datastore-storage-adapter@2.1.89...@aws-amplify/datastore-storage-adapter@2.2.0) (2025-09-29)
 
 ### Features